move clients_names function to clients data file and update paths
