+++ frontend/pages/data/data.py
@@ -25,16 +25,6 @@ def username() -> List[Select]:
     return username_rows
 
 
-def clients_names() -> List[Select]:
-    api_client_name = f"{base_url}/api/clients/active"
-    response_client_name = requests.get(api_client_name)
-    client_name_rows = [Select(value="", display_value="select client")]
-    for item in response_client_name.json():
-        d = Select(value=item["id"], display_value=item["name"])
-        client_name_rows.append(d)
-    return client_name_rows
-
-
 def client_name_by_epic_id(epic_id) -> Select:
     api_client_name_id = f"{base_url}/api/epics/{epic_id}/client-name"
     response_client_name_id = requests.get(api_client_name_id)
